Move ValidationError to separate file and make CommandResponse package private
