+++ packages/cqrs/lib/cqrs.dart
@@ -64,8 +64,8 @@
 /// code contract generator.
 library;
 
-export 'src/command_result.dart';
 export 'src/cqrs.dart';
 export 'src/cqrs_error.dart';
 export 'src/cqrs_result.dart';
 export 'src/transport_types.dart';
+export 'src/validation_error.dart';
