+++ packages/cqrs/test/command_result_test.dart
@@ -8,7 +8,7 @@ void main() {
 
     group('fields values are correct', () {
       test('when constructed without errors', () {
-        const result = CommandResult([]);
+        const result = CommandResponse([]);
 
         expect(result.success, true);
         expect(result.failed, false);
@@ -16,7 +16,7 @@ void main() {
       });
 
       test('when constructed with errors', () {
-        const result = CommandResult([error1, error2]);
+        const result = CommandResponse([error1, error2]);
 
         expect(result.success, false);
         expect(result.failed, true);
@@ -24,7 +24,7 @@ void main() {
       });
 
       test('when constructed with success constructor', () {
-        const result = CommandResult.success();
+        const result = CommandResponse.success();
 
         expect(result.success, true);
         expect(result.failed, false);
@@ -32,7 +32,7 @@ void main() {
       });
 
       test('when constructed with success constructor', () {
-        final result = CommandResult.failed([error1]);
+        final result = CommandResponse.failed([error1]);
 
         expect(result.success, false);
         expect(result.failed, true);
@@ -42,7 +42,7 @@ void main() {
 
     group('hasError returns correct values', () {
       test('when there are some errors present', () {
-        const result = CommandResult([error1, error2]);
+        const result = CommandResponse([error1, error2]);
 
         expect(result.hasError(1), true);
         expect(result.hasError(2), true);
@@ -50,7 +50,7 @@ void main() {
       });
 
       test('when there are no errors present', () {
-        const result = CommandResult([]);
+        const result = CommandResponse([]);
 
         expect(result.hasError(1), false);
         expect(result.hasError(2), false);
@@ -60,21 +60,21 @@ void main() {
 
     group('hasErrorForProperty returns correct values', () {
       test('when there are some errors present', () {
-        const result = CommandResult([error1, error2]);
+        const result = CommandResponse([error1, error2]);
 
         expect(result.hasErrorForProperty(1, 'Property1'), true);
         expect(result.hasErrorForProperty(2, 'Property2'), true);
       });
 
       test('when there are errors but for different properties', () {
-        const result = CommandResult([error1, error2]);
+        const result = CommandResponse([error1, error2]);
 
         expect(result.hasErrorForProperty(1, 'Property2'), false);
         expect(result.hasErrorForProperty(2, 'Property1'), false);
       });
 
       test('when there are no errors present', () {
-        const result = CommandResult([]);
+        const result = CommandResponse([]);
 
         expect(result.hasErrorForProperty(1, 'Property1'), false);
         expect(result.hasErrorForProperty(2, 'Property1'), false);
@@ -85,7 +85,7 @@ void main() {
 
     group('is correctly deserialized from JSON', () {
       test('with some validation errors', () {
-        final result = CommandResult.fromJson(<String, dynamic>{
+        final result = CommandResponse.fromJson(<String, dynamic>{
           'WasSuccessful': false,
           'ValidationErrors': [
             {
@@ -98,7 +98,7 @@ void main() {
 
         expect(
           result,
-          isA<CommandResult>()
+          isA<CommandResponse>()
               .having((r) => r.success, 'success', false)
               .having(
                 (r) => r.errors,
@@ -118,14 +118,14 @@ void main() {
       });
 
       test('without validation errors, with success', () {
-        final result = CommandResult.fromJson(<String, dynamic>{
+        final result = CommandResponse.fromJson(<String, dynamic>{
           'WasSuccessful': true,
           'ValidationErrors': <Map<String, dynamic>>[],
         });
 
         expect(
           result,
-          isA<CommandResult>()
+          isA<CommandResponse>()
               .having((r) => r.success, 'success', true)
               .having((r) => r.errors, 'errors', isEmpty),
         );
@@ -134,7 +134,7 @@ void main() {
 
     group('is correctly serialized to JSON', () {
       test('with some validation errors', () {
-        final json = CommandResult.fromJson(<String, dynamic>{
+        final json = CommandResponse.fromJson(<String, dynamic>{
           'WasSuccessful': false,
           'ValidationErrors': [
             {
@@ -162,7 +162,7 @@ void main() {
       });
 
       test('without validation errors, with success', () {
-        final json = CommandResult.fromJson(<String, dynamic>{
+        final json = CommandResponse.fromJson(<String, dynamic>{
           'WasSuccessful': true,
           'ValidationErrors': <Map<String, dynamic>>[],
         }).toJson();
@@ -174,13 +174,13 @@ void main() {
 
     group('returns correct string value', () {
       test('for success', () {
-        const error = CommandResult.success();
+        const error = CommandResponse.success();
 
         expect(error.toString(), 'CommandResult([])');
       });
 
       test('for errors', () {
-        final error = CommandResult.failed(const [
+        final error = CommandResponse.failed(const [
           ValidationError(23, 'message', 'propertyName'),
           ValidationError(29, 'message2', 'propertyName2'),
         ]);
